requirements.txt
streamlit==1.29.0
boto3==1.34.0
python-pptx==0.6.21